Database.cpp: Fix compiler warning with cast. CtrlChessTxt.cpp: On paste text, transform non ascii to ? rather than simply clearing bit 7, handle Unicode en-dash, we need to take a close look at Unicode issues

+++ src/Database.cpp
@@ -551,7 +551,7 @@ int Database::LoadPlayerGamesWithQuery(  std::string &player_name, bool white, s
 
 int Database::GetRow( int row, CompactGame *pact )
 {
-    if( 0<=row && row<tiny_db.in_memory_game_cache.size() )
+    if( 0<=row && static_cast<unsigned int>(row)<tiny_db.in_memory_game_cache.size() )
     {
         smart_ptr<ListableGame> p = tiny_db.in_memory_game_cache[row];
         p->GetCompactGame(*pact);
